refactor: . move parseAndValidateRecurrence() to helpers file
