+++ src/ui/RecurrenceEditor.svelte
@@ -1,39 +1,13 @@
 <script lang="ts">
     import { TASK_FORMATS } from '../Config/Settings';
-    import { Recurrence } from '../Task/Recurrence';
     import type { EditableTask } from './EditableTask';
-    import { labelContentWithAccessKey } from './EditTaskHelpers';
+    import { labelContentWithAccessKey, parseAndValidateRecurrence } from './EditTaskHelpers';
 
     export let editableTask: EditableTask;
     export let isRecurrenceValid: boolean;
     export let accesskey: string | null;
 
     let parsedRecurrence: string;
-
-    function parseAndValidateRecurrence(editableTask: EditableTask) {
-        // NEW_TASK_FIELD_EDIT_REQUIRED
-        if (!editableTask.recurrenceRule) {
-            return { parsedRecurrence: '<i>not recurring</>', isRecurrenceValid: true };
-        }
-
-        const recurrenceFromText = Recurrence.fromText({
-            recurrenceRuleText: editableTask.recurrenceRule,
-            // Only for representation in the modal, no dates required.
-            startDate: null,
-            scheduledDate: null,
-            dueDate: null,
-        })?.toText();
-
-        if (!recurrenceFromText) {
-            return { parsedRecurrence: '<i>invalid recurrence rule</i>', isRecurrenceValid: false };
-        }
-
-        if (editableTask.startDate || editableTask.scheduledDate || editableTask.dueDate) {
-            return { parsedRecurrence: recurrenceFromText, isRecurrenceValid: true };
-        }
-
-        return { parsedRecurrence: '<i>due, scheduled or start date required</i>', isRecurrenceValid: false };
-    }
 
     $: ({ parsedRecurrence, isRecurrenceValid } = parseAndValidateRecurrence(editableTask));
